refactor(18846): clean panels
