+++ hivemq-edge/src/frontend/src/extensions/datahub/components/panels/BehaviorPolicyPanel.tsx
@@ -2,10 +2,10 @@ import { FC, useMemo } from 'react'
 import { Node } from 'reactflow'
 import { Card, CardBody } from '@chakra-ui/react'
 
+import { MOCK_BEHAVIOR_POLICY_SCHEMA } from '../../api/specs/'
 import useDataHubDraftStore from '../../hooks/useDataHubDraftStore.ts'
 import { BehaviorPolicyData, PanelProps } from '../../types.ts'
 import { ReactFlowSchemaForm } from '../helpers/ReactFlowSchemaForm.tsx'
-import { MOCK_VBEHAVIOR_POLICY_SCHEMA } from '@/extensions/datahub/api/specs/BehaviorPolicyData.ts'
 
 export const BehaviorPolicyPanel: FC<PanelProps> = ({ selectedNode, onFormSubmit }) => {
   const { nodes } = useDataHubDraftStore()
@@ -19,12 +19,10 @@ export const BehaviorPolicyPanel: FC<PanelProps> = ({ selectedNode, onFormSubmit
     <Card>
       <CardBody>
         <ReactFlowSchemaForm
-          schema={MOCK_VBEHAVIOR_POLICY_SCHEMA.schema}
+          schema={MOCK_BEHAVIOR_POLICY_SCHEMA.schema}
           // uiSchema={MOCK_TOPIC_FILTER_SCHEMA.uiSchema}
           formData={data}
           onSubmit={onFormSubmit}
-          onChange={() => console.log('changed')}
-          onError={() => console.log('errors')}
         />
       </CardBody>
     </Card>
