+++ hivemq-edge/src/frontend/src/extensions/datahub/components/panels/SchemaPanel.tsx
@@ -7,9 +7,9 @@ import { CustomValidator } from '@rjsf/utils'
 // import { customizeValidator } from '@rjsf/validator-ajv8'
 // import { parse } from 'protobufjs'
 
+import { MOCK_SCHEMA_SCHEMA } from '../../api/specs/'
 import { PanelProps, SchemaData } from '../../types.ts'
 import useDataHubDraftStore from '../../hooks/useDataHubDraftStore.ts'
-import { MOCK_SCHEMA_SCHEMA } from '../../api/specs/SchemaData.ts'
 import { ReactFlowSchemaForm, datahubRJSFWidgets } from '../helpers'
 
 export const SchemaPanel: FC<PanelProps> = ({ selectedNode, onFormSubmit }) => {
@@ -64,8 +64,6 @@ export const SchemaPanel: FC<PanelProps> = ({ selectedNode, onFormSubmit }) => {
           widgets={datahubRJSFWidgets}
           customValidate={customValidate}
           onSubmit={onFormSubmit}
-          onChange={() => console.log('changed')}
-          onError={() => console.log('errors')}
         />
       </CardBody>
     </Card>
